model

+++ src/main/java/com/sk/travel/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public
-interface UserRepository extends JpaRepository<User, Integer>{
+interface UserRepository extends JpaRepository<User, Long>{
 //    User findByPhonenumber(String phonenumber);
     User findByEmail(String email);
